Fix indexing of deep objects
JsonObject.getMap returns the original map from inside JsonObject. Deep nested
objects are contained there as JsonObject Objects which implements Iterable of
Map.Entry. IndexRequestBuilder.setSource(Map) uses XContentBuilder to transform
the objects it gets. For an Iterable it creates an array of the items, in that
case their toString value. This basically destroys indexing of nested object
with the module, becuase the become an array of "key=value" Strings.

This patch changes the "getMap" to "encode" in order to hand the actual json to
Elasticsearch.